Still trying to find the best architecture for this; using inheritance seemed like a smell to me so I've moved towards a composition-based design.
Essentially, the model object (PTProject in this case) is your model; how it gets its data (either from a dictionary pulled and parsed from the API, or an NSManagedObject) is an implementation detail. It retains a reference to its NSManagedObject if it has one but NSManagedObject is nothing more than a dumb data store and interface to CoreData here. 

I started off doing the synching in the controller but quickly pulled this out into the PTSyncManager class.

You'll notice I'm trying to avoid any kind of singleton (CoreDataManager aside, but I'm thinking about de-singletonizing that too). We should declare app-wide dependencies like the CoreData stack and PTSyncManager in the app delegate but pass them to objects that need them using accessor-injection.

The rest of this is commit is throwing the whole UI into a navigation controller so I can add a refresh button.

Big commit message? That will teach me not to commit in a more atomic fashion.