Refactored TrackerViewController to use an NSFetchedResultsController.
This means we can remove all of the query stuff from PTProject and leave it
in the controller layer where it belongs (much of which is abstracted away
by NSFetchedResultsController).

The only issue is that NSFetchedResultsController deals in NSManagedObjects
but we want to use our model object, so we need to wrap each NSManagedObject
that we pull out of the fetched results in our own object.